refactor: move helpers out of vm_helper.go and add MIContainers label
- Move TestHTTPSConnectivity to connectivity_helper.go (network reachability)
- Move GetPrivateKASInternalIP to resource_group_helper.go (Azure RG query)
- Add required MIContainers(1) label to private KAS test
